refactor(timesfm): specialize family implementation
Retain only TimesFM configuration, checkpoint, graph, and time-series operations while preserving current normalization and reduced-precision behavior.
